requirements.txt
aiofiles==25.1.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
exceptiongroup==1.3.0
fastjsonschema==2.21.2
h11==0.16.0
html5tagger==1.3.0
httptools==0.7.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
kagglehub==0.3.13
lxml==6.0.2
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
reactpy==1.1.0
requests==2.32.5
sanic==25.3.0
Sanic-Cors==2.2.0
sanic-routing==23.12.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
tracerite==1.1.3
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
